다른 클래스에서 사용할 수 있도록 GameState,PlayerState 내용 첨부(struct 기반)

제한시간 UI 1초마다 이벤트 발생 -> UI 업데이트 (GameMode -> GameState(Event) -> StagePlayerController -> Widget)

기존 System 폴더 내 리펙토링 파일 주석처리 한거 지움

dev 최신화 된 내용으로 진행한거라 따로 dev -> System merge 하지 않았음 (필요 X)

+++ Source/ST/Public/System/StageInfoData.h
@@ -14,6 +14,6 @@ struct FStageInfoRow : public FTableRowBase
 	FString StageName;
 
 	UPROPERTY( EditAnywhere, BlueprintReadWrite )
-	float TimeLimit;
+	int32 TimeLimit;
 	
 };
